Removed custom functions when setting attributes
Specifying a function doesn't work with the current design: the custom function will apply when an attribute is set *multiple times within a single list literal* but will *not* apply when sets are combined with `<>`, which would be confusing behavior.

More generally, I'm realizing that I need a clearer conceptual design for attribute sets—maybe even splitting into separate types for sets and diffs of sets.